Further refactor.
Moved the cfn resource identifiers to a yaml config file - it shouldnt
be hardcoded into a dictionary - easier to update / modify in a yaml
config file.

added update_stack and waiter to aws_handler

slowly going through and refactoring the rest of index.py - before
adding in the functionality I need to it.

	new file:   configs/cloudformation.yaml
	modified:   configs/config.yaml
	modified:   libs/aws_handler.py
	modified:   libs/command_checker.py
	modified:   stack_rename